Updated environment and model with more parameters

+++ envs/yahoofinance_env.py
@@ -32,12 +32,14 @@ class YahooFinanceEnv(gym.Env):
         self.inventory = []
 
     def step(self, action):
+        if self.current_step >= self.price_history.get_length():
+            return [[0,0,0,0,0,0]], 0, True, {}
+
         reward = self._take_action(action)
 
         close_price = self.trading_days[self.current_step].get_close()
         done = True if self.current_step >= len(self.price_history.get_dataframe()) or (getTotalPriceOfHoldings(self.inventory, close_price) == 0 and (self.cash * self.risk) < close_price) else False
 
-        #plt.scatter(self.price_history.get_dataframe().iloc[self.current_step]['time'], self.balance, marker='+', label='Balance', color='blue')
         self.current_step += 1
 
         observation = self._get_observation(self.current_step)
@@ -57,50 +59,24 @@ class YahooFinanceEnv(gym.Env):
 
     def _get_observation(self,step):
         observations = []
+
         for i in range(self.current_step - self._lookbabck, self.current_step):
             candle = self.trading_days[i].get_candle()
-            obs = [
-                candle.get_direction(),
-                candle.get_size(),
-                candle.get_wick_down(),
-                candle.get_wick_up()
-            ]
-            observations.append(obs)
+            observations.append(candle.to_vector())
+
         observations = np.array(observations)
-        observations.reshape(self._lookbabck, 1, 4)
         return observations
 
     def render(self, mode='human', close=False):
+        if self.current_step >= self.price_history.get_length():
+            return
+
         if close:
             if self.viewer is not None:
                 os.kill(self.viewer.pid, signal.SIGKILL)
         else:
             close_price = self.trading_days[self.current_step].get_close()
             print(self.trading_days[self.current_step].get_date().strftime('%m/%d/%Y') + " Cash " + formatPrice(self.cash) + " | Holdings: " + formatPriceOfHoldings(self.inventory, close_price) + " | Balance: " + formatPrice(self.balance))
-
-    # def _start_viewer(self):
-    #     mondays = WeekdayLocator(MONDAY)        # major ticks on the mondays
-    #     alldays = DayLocator()              # minor ticks on the days
-    #     weekFormatter = DateFormatter('%b %d')  # e.g., Jan 12
-    #     dayFormatter = DateFormatter('%d')      # e.g., 12
-    #     self.fig, self.ax = plt.subplots()
-    #     self.fig.subplots_adjust(bottom=0.2)
-    #     self.ax.xaxis.set_major_locator(mondays)
-    #     self.ax.xaxis.set_minor_locator(alldays)
-    #     self.ax.xaxis.set_major_formatter(weekFormatter)
-    #
-    #     candlestick_ohlc(self.ax, self.price_history.get_dataframe().values, width=0.6)
-    #
-    #     self.ax.xaxis_date()
-    #     self.ax.autoscale_view()
-    #     plt.setp(plt.gca().get_xticklabels(), rotation=45, horizontalalignment='right')
-    #
-    #     for i in range(len(self.price_history.get_dataframe())):
-    #         y = np.random.random()
-    #         plt.scatter(self.price_history.get_dataframe().iloc[i]['time'], y, marker='+', label='Balance', color='blue')
-    #         plt.pause(0.05)
-    #
-    #     plt.show()
 
     def _take_action(self, action):
         close_price = self.trading_days[self.current_step].get_close()
@@ -111,7 +87,6 @@ class YahooFinanceEnv(gym.Env):
             cost = close_price * shares
             self.inventory.append([shares, cost])
             self.cash = self.cash - cost - self.trading_fee
-            #reward -= self.trading_fee / 100 #trading fee
             #print("Buy " + str(shares) + " shares @ " + formatPrice(close_price))
 
         elif action == SELL and len(self.inventory) > 0:  # sell
@@ -120,7 +95,7 @@ class YahooFinanceEnv(gym.Env):
             bought_shares = order[0]
             current_value = bought_shares * close_price
             self.cash += current_value - self.trading_fee
-            take = (current_value - bought_price) / bought_shares
+            take = (current_value - bought_price)
             reward = math.tanh(take)
             #print("Sell " + str(bought_shares) + " @ " + formatPrice(close_price) + " [Net " + formatPrice(current_value - bought_price) + "]")
